remove dependency on github.com/hpcloud/tail by using io.TeeReader
